requirements.txt
aiohttp
beautifulsoup4
PyYAML
python-slugify
tqdm
ijson